Refactor consumer group.
