+++ lib/consumerGroup.js
@@ -2,19 +2,14 @@
 
 var debug = require('debug')('kafka-node:ConsumerGroup');
 var util = require('util');
-var retry = require('retry');
 var highLevelConsumer = require('./highLevelConsumer');
 var Client = require('./client');
 var _ = require('lodash');
 var async = require('async');
 var validateConfig = require('./utils').validateConfig;
+var ConsumerGroupRecovery = require('./ConsumerGroupRecovery');
 var createTopicPartitionList = require('./utils').createTopicPartitionList;
-var GroupCoordinatorNotAvailable = require('./errors/GroupCoordinatorNotAvailableError');
-var NotCoordinatorForGroup = require('./errors/NotCoordinatorForGroupError');
-var IllegalGeneration = require('./errors/IllegalGenerationError');
-var GroupLoadInProgress = require('./errors/GroupLoadInProgressError');
-var UnknownMemberId = require('./errors/UnknownMemberIdError');
-var RebalanceInProgress = require('./errors/RebalanceInProgressError');
+
 var assert = require('assert');
 var roundRobin = require('./assignment/roundrobin');
 
@@ -48,12 +43,14 @@ function ConsumerGroup (memberOptions, topics) {
   this.topics = topics;
   this.options.protocol.subscription = topics || [];
 
+  this.recovery = new ConsumerGroupRecovery(this);
+
   if (this.options.connectOnReady) {
     this.client.once('ready', this.connect.bind(this));
   }
 
   this.client.on('error', function (err) {
-    debug('Error from %j', err);
+    debug('Error from %s', self.client.clientId, err);
     self.emit('error', err);
   });
 
@@ -116,6 +113,52 @@ function mapTopicToPartitions (metadata) {
   return _.mapValues(metadata, Object.keys);
 }
 
+ConsumerGroup.prototype.handleJoinGroup = function (joinGroupResponse, callback) {
+  debug('joinGroupResponse %j from %s', joinGroupResponse, this.client.clientId);
+
+  this.isLeader = (joinGroupResponse.leaderId === joinGroupResponse.memberId);
+  this.generationId = joinGroupResponse.generationId;
+  this.memberId = joinGroupResponse.memberId;
+
+  var groupAssignment;
+  if (this.isLeader) {
+    // assign partitions
+    return this.assignPartitions(joinGroupResponse.groupProtocol, joinGroupResponse.members, callback);
+  }
+  callback(null, groupAssignment);
+};
+
+ConsumerGroup.prototype.handleSyncGroup = function (syncGroupResponse, callback) {
+  debug('SyncGroup Response');
+  var self = this;
+  var ownedTopics = Object.keys(syncGroupResponse.partitions);
+  if (ownedTopics.length) {
+    debug('%s owns topics: ', self.client.clientId, syncGroupResponse.partitions);
+
+    self.fetchOffset(syncGroupResponse.partitions, function (error, offsets) {
+      if (error) {
+        return callback(error);
+      }
+
+      debug('%s fetchOffset Response: %j', self.client.clientId, offsets);
+
+      self.topicPayloads = [];
+
+      var payloads = self.buildPayloads(createTopicPartitionList(syncGroupResponse.partitions));
+      payloads.forEach(function (p) {
+        var offset = offsets[p.topic][p.partition];
+        if (offset === -1) offset = 0;
+        p.offset = offset;
+        self.topicPayloads.push(p);
+      });
+
+      callback(null, true);
+    });
+  } else { // no partitions assigned
+    callback(null, false);
+  }
+};
+
 ConsumerGroup.prototype.connect = function () {
   if (this.connecting) {
     debug('Connect ignored. Currently connecting.');
@@ -144,18 +187,7 @@ ConsumerGroup.prototype.connect = function () {
     },
 
     function (joinGroupResponse, callback) {
-      debug('joinGroupResponse %j from %s', joinGroupResponse, self.client.clientId);
-
-      self.isLeader = (joinGroupResponse.leaderId === joinGroupResponse.memberId);
-      self.generationId = joinGroupResponse.generationId;
-      self.memberId = joinGroupResponse.memberId;
-
-      var groupAssignment;
-      if (self.isLeader) {
-        // assign partitions
-        return self.assignPartitions(joinGroupResponse.groupProtocol, joinGroupResponse.members, callback);
-      }
-      callback(null, groupAssignment);
+      self.handleJoinGroup(joinGroupResponse, callback);
     },
 
     function (groupAssignment, callback) {
@@ -164,43 +196,16 @@ ConsumerGroup.prototype.connect = function () {
     },
 
     function (syncGroupResponse, callback) {
-      debug('SyncGroup Response');
-      var ownedTopics = Object.keys(syncGroupResponse.partitions);
-      if (ownedTopics.length) {
-        debug('%s owns topics: ', self.client.clientId, syncGroupResponse.partitions);
-
-        self.fetchOffset(syncGroupResponse.partitions, function (error, offsets) {
-          if (error) {
-            return callback(error);
-          }
-
-          debug('%s fetchOffset Response: %j', self.client.clientId, offsets);
-
-          self.topicPayloads = [];
-
-          var payloads = self.buildPayloads(createTopicPartitionList(syncGroupResponse.partitions));
-
-          payloads.forEach(function (p) {
-            var offset = offsets[p.topic][p.partition];
-            if (offset === -1) offset = 0;
-            p.offset = offset;
-            self.topicPayloads.push(p);
-          });
-
-          callback(null, true);
-        });
-      } else { // no partitions assigned
-        callback(null, false);
-      }
+      self.handleSyncGroup(syncGroupResponse, callback);
     }
   ], function (error, startFetch) {
     self.connecting = false;
     if (error) {
-      return self.tryToRecoverFrom(error, 'connect');
+      return self.recovery.tryToRecoverFrom(error, 'connect');
     }
 
     self.ready = true;
-    self.lastError = null;
+    self.recovery.clearError();
 
     debug('generationId', self.generationId);
 
@@ -210,77 +215,6 @@ ConsumerGroup.prototype.connect = function () {
     self.emit('connect');
     self.startHearbeats();
   });
-};
-
-var recoverableErrors = [
-  {
-    errors: [GroupCoordinatorNotAvailable, IllegalGeneration, GroupLoadInProgress, RebalanceInProgress]
-  },
-  {
-    errors: [NotCoordinatorForGroup],
-    handler: function () {
-      this.client.coordinatorId = null;
-    }
-  },
-  {
-    errors: [UnknownMemberId],
-    handler: function () {
-      this.memberId = null;
-    }
-  }
-];
-
-function isErrorInstanceOf (error, errors) {
-  return errors.some(function (errorClass) {
-    return error instanceof errorClass;
-  });
-}
-
-ConsumerGroup.prototype.tryToRecoverFrom = function (error, source) {
-  this.ready = false;
-  this.stopHeartbeats();
-
-  var retryTimeout = false;
-  var retry = recoverableErrors.some(function (recoverableItem) {
-    if (isErrorInstanceOf(error, recoverableItem.errors)) {
-      recoverableItem.handler && recoverableItem.handler.call(this, error);
-      return true;
-    }
-    return false;
-  }, this);
-
-  if (retry) {
-    retryTimeout = this.getRetryTimeout(error);
-  }
-
-  if (retry && retryTimeout) {
-    debug('RECOVERY from %s: %s retrying', source, this.memberId || this.client.clientId, error);
-    this.scheduleReconnect(retryTimeout);
-  } else {
-    this.emit('error', error);
-  }
-  this.lastError = error;
-};
-
-ConsumerGroup.prototype.getRetryTimeout = function (error) {
-  assert(error);
-  if (!this._timeouts) {
-    this._timeouts = retry.timeouts({
-      retries: this.options.retries,
-      factor: this.options.retryFactor
-    });
-  }
-
-  if (this._retryIndex == null || this.lastError == null ||
-      error.errorCode !== this.lastError.errorCode) {
-    this._retryIndex = 0;
-  }
-
-  var index = this._retryIndex++;
-  if (index >= this._timeouts.length) {
-    return false;
-  }
-  return this._timeouts[index];
 };
 
 ConsumerGroup.prototype.scheduleReconnect = function (timeout) {
