+++ lib/highLevelConsumer.js
@@ -623,7 +623,7 @@ HighLevelConsumer.prototype.removeTopics = function (topics, cb) {
   this.client.removeTopicMetadata(topics, cb);
 };
 
-HighLevelConsumer.prototype._leaveGroup = function (cb) {
+HighLevelConsumer.prototype.leaveGroup = function (cb) {
   var self = this;
   async.parallel([
     function (callback) {
@@ -652,7 +652,7 @@ HighLevelConsumer.prototype.close = function (force, cb) {
 
   async.series([
     function (callback) {
-      self._leaveGroup(callback);
+      self.leaveGroup(callback);
     },
     function (callback) {
       if (force) {
